requirements.txt
streamlit
pytesseract
Pillow
spacy
transformers
torch
pandas
